Split fuzz tooling layers
